Flashcard
It is a web project to learn something like a game. There are few number of card that contain country name and there capital on the backside of the card. As learner will point the curser on the card it will show the capital of the country that is in the card by using rotation.